LibJS: Make DataView::byte_offset() return u32
This fixes structured serialization of DataView. It was expected
to be uniform with TypedArray, which returns u32 for byte_offset().

This was covered by a number of WPT infrastructure tests, which this
commit also imports.

+++ Userland/Libraries/LibJS/Runtime/DataView.h
@@ -24,7 +24,7 @@ public:
 
     ArrayBuffer* viewed_array_buffer() const { return m_viewed_array_buffer; }
     ByteLength const& byte_length() const { return m_byte_length; }
-    size_t byte_offset() const { return m_byte_offset; }
+    u32 byte_offset() const { return m_byte_offset; }
 
 private:
     DataView(ArrayBuffer*, ByteLength byte_length, size_t byte_offset, Object& prototype);
